docs/AI/Context_Graph/requirements.txt
neo4j>=5.15.0
openai>=1.3.0
networkx>=3.1
matplotlib>=3.7.0
python-dotenv>=1.0.0
